requirements.txt
PyPDF2
pandas
pdfminer.six